improve tests with in memory blockchains and admins
