+++ package.json
@@ -25,7 +25,9 @@
     "jsonwebtoken": "^9.0.0",
     "jwt-decode": "^3.1.2",
     "mongodb": "^5.1.0",
+    "mongodb-memory-server": "^8.12.2",
     "puppeteer": "^19.9.0",
+    "sinon": "^15.0.4",
     "tough-cookie": "^4.1.2"
   },
   "devDependencies": {
