+++ src/routes/admins.js
@@ -1,11 +1,12 @@
 import express from 'express';
-import getDBConnection from '../db/conn.js';
+import { Database } from '../db/conn.js';
 import isRequired from '../utils/auth-utils.js';
 
 const router = express.Router();
 
 router.get('/', isRequired, async (req, res) => {
-  const collection = (await getDBConnection(req)).collection('admins');
+  const db = await Database.getInstance(req);
+  const collection = db.collection('admins');
   res.status(200).send(
     (await collection.findOne({
       userId: { $regex: res.locals.userId, $options: 'i' },
