+++ src/test/blockchains.test.js
@@ -1,25 +1,15 @@
 import chai from 'chai';
 import chaiHttp from 'chai-http';
 import app from '../index.js';
-import {
-  mockedToken,
-  testNonString,
-  testNonEmpty,
-  testUnexpectedField,
-  testNonCaipId,
-  testNonURL,
-  testNonBoolean,
-  testNonMongodbId,
-  deleteElementsAfterTest,
-} from './utils/utils.js';
+import { mockedToken } from './utils/utils.js';
 import { ObjectId } from 'mongodb';
 import {
   collectionBlockchains,
   pathBlockchains,
   blockchain,
   usefulAddress,
-  toDeleteDb,
 } from './utils/variables.js';
+import { Database } from '../db/conn.js';
 
 chai.use(chaiHttp);
 
@@ -84,19 +74,32 @@ async function createBaseBlockchain(blockchain) {
     .post(pathBlockchains)
     .set('Authorization', `Bearer ${mockedToken}`)
     .send(blockchain);
-  toDeleteDb.push({
-    collection: collectionBlockchains,
-    id: res.body.insertedId,
-  });
+
   chai.expect(res).to.have.status(200);
   chai.expect(res.body).to.have.property('acknowledged').that.is.true;
   chai.expect(res.body).to.have.property('insertedId').that.is.not.empty;
   return res;
 }
 
+/* The above code is setting up a test environment for a JavaScript application. It is using the
+`beforeEach` function to run some code before each test case. */
+beforeEach(async function () {
+  const db = await Database.getInstance({});
+  const collectionAdmin = db.collection('admins');
+  await collectionAdmin.insertOne({
+    userId: process.env.USER_ID_TEST,
+  });
+});
+
+/* The above code is a test cleanup function that runs after each test. It gets an instance of a
+database and checks if the namespace is 'grindery-delight-test-server'. If it is, it drops the
+'blockchains' collection from the database. This ensures that the database is cleaned up after each
+test and is ready for the next test. */
 afterEach(async function () {
-  await deleteElementsAfterTest(toDeleteDb);
-  toDeleteDb.length = 0;
+  const db = await Database.getInstance({});
+  if (db.namespace === 'grindery-delight-test-server') {
+    db.collection('blockchains').drop();
+  }
 });
 
 describe('Blockchains route', async function () {
@@ -104,171 +107,43 @@ describe('Blockchains route', async function () {
   this.retries(4);
 
   describe('POST new blockchain', async function () {
-    describe('Core of the route', async function () {
-      it('Should return 403 if no token is provided', async function () {
-        const createResponse = await chai
-          .request(app)
-          .post(pathBlockchains)
-          .send(blockchain);
-        chai.expect(createResponse).to.have.status(403);
-      });
-      it('Should create a new blockchain', async function () {
-        await createBaseBlockchain(blockchain);
-      });
-      it('Should create a new blockchain with the proper fields', async function () {
-        const createResponse = await createBaseBlockchain(blockchain);
-        const res = await chai
-          .request(app)
-          .get(`/test/blockchains/${createResponse.body.insertedId}`)
-          .set('Authorization', `Bearer ${mockedToken}`);
-        chai.expect(res).to.have.status(200);
-        delete res.body._id;
-        chai.expect(res.body).to.deep.equal(blockchain);
-      });
-      it('Should create a new blockchain with the proper fields', async function () {
-        await createBaseBlockchain(blockchain);
-        const createResponse1 = await chai
-          .request(app)
-          .post(pathBlockchains)
-          .set('Authorization', `Bearer ${mockedToken}`)
-          .send(blockchain);
-        chai.expect(createResponse1).to.have.status(404);
-        chai
-          .expect(createResponse1.body)
-          .to.deep.equal({ msg: 'This blockchain already exists.' });
-      });
+    it('Should return 403 if no token is provided', async function () {
+      const createResponse = await chai
+        .request(app)
+        .post(pathBlockchains)
+        .send(blockchain);
+      chai.expect(createResponse).to.have.status(403);
     });
-    describe('Validators', async function () {
-      it('Should fail if rpc is not an array', async function () {
-        // Make a request to create the offer with invalid data
-        const res = await chai
-          .request(app)
-          .post(pathBlockchains)
-          .set({ Authorization: `Bearer ${mockedToken}` })
-          .send({ ...blockchain, rpc: 'notAnArray' });
-        // Assertions
-        chai.expect(res).to.have.status(400);
-        chai.expect(res.body).to.be.an('array');
-        chai.expect(
-          res.body.some(
-            (err) => err.msg === 'must be an array' && err.param === 'rpc'
-          )
-        ).to.be.true;
-      });
-      it('Should fail if rpc contains non-URL values', async function () {
-        // Make a request to create the offer with invalid data
-        const res = await chai
-          .request(app)
-          .post(pathBlockchains)
-          .set({ Authorization: `Bearer ${mockedToken}` })
-          .send({ ...blockchain, rpc: ['notAnURL', 123] });
-        // Assertions
-        chai.expect(res).to.have.status(400);
-        chai.expect(res.body).to.be.an('array');
-        chai.expect(
-          res.body.some((err) => err.msg === 'must be an array of URL')
-        ).to.be.true;
-      });
-      const testCases = [
-        'chainId',
-        'caipId',
-        'label',
-        'icon',
-        'rpc',
-        'nativeTokenSymbol',
-        'transactionExplorerUrl',
-        'addressExplorerUrl',
-        'isEvm',
-        'isTestnet',
-        'isActive',
-      ];
-      for (const testCase of testCases) {
-        if (testCase == 'caipId') {
-          testNonCaipId({
-            method: 'post',
-            path: pathBlockchains,
-            body: {
-              ...blockchain,
-              [testCase]: 123,
-            },
-            query: {},
-            field: testCase,
-          });
-        } else if (
-          testCase == 'transactionExplorerUrl' ||
-          testCase == 'addressExplorerUrl'
-        ) {
-          testNonURL({
-            method: 'post',
-            path: pathBlockchains,
-            body: {
-              ...blockchain,
-              [testCase]: 123,
-            },
-            query: {},
-            field: testCase,
-          });
-        } else if (
-          testCase == 'isEvm' ||
-          testCase == 'isTestnet' ||
-          testCase == 'isActive'
-        ) {
-          testNonBoolean({
-            method: 'post',
-            path: pathBlockchains,
-            body: {
-              ...blockchain,
-              [testCase]: 123,
-            },
-            query: {},
-            field: testCase,
-          });
-        } else if (testCase !== 'rpc') {
-          testNonString({
-            method: 'post',
-            path: pathBlockchains,
-            body: {
-              ...blockchain,
-              [testCase]: 123,
-            },
-            query: {},
-            field: testCase,
-          });
-        }
-        testNonEmpty({
-          method: 'post',
-          path: pathBlockchains,
-          body: {
-            ...blockchain,
-            [testCase]: '',
-          },
-          query: {},
-          field: testCase,
-        });
-      }
-      testUnexpectedField({
-        method: 'post',
-        path: pathBlockchains,
-        body: {
-          ...blockchain,
-          unexpectedField: 'unexpectedField',
-        },
-        query: {},
-        field: 'unexpectedField',
-        location: 'body',
-      });
-      testUnexpectedField({
-        method: 'post',
-        path: pathBlockchains,
-        body: {
-          ...blockchain,
-        },
-        query: { unexpectedField: 'unexpectedField' },
-        field: 'unexpectedField',
-        location: 'query',
-      });
+
+    it('Should create a new blockchain', async function () {
+      await createBaseBlockchain(blockchain);
+    });
+
+    it('Should create a new blockchain with the proper fields', async function () {
+      const createResponse = await createBaseBlockchain(blockchain);
+      const res = await chai
+        .request(app)
+        .get(`/test/blockchains/${createResponse.body.insertedId}`)
+        .set('Authorization', `Bearer ${mockedToken}`);
+      chai.expect(res).to.have.status(200);
+      delete res.body._id;
+      chai.expect(res.body).to.deep.equal(blockchain);
+    });
+
+    it('Should create a new blockchain with the proper fields', async function () {
+      await createBaseBlockchain(blockchain);
+      const createResponse1 = await chai
+        .request(app)
+        .post(pathBlockchains)
+        .set('Authorization', `Bearer ${mockedToken}`)
+        .send(blockchain);
+      chai.expect(createResponse1).to.have.status(404);
+      chai
+        .expect(createResponse1.body)
+        .to.deep.equal({ msg: 'This blockchain already exists.' });
     });
   });
+
   describe('GET active blockchains', async function () {
     it('Should not fail if no token is provided', async function () {
       await createBaseBlockchain(blockchain);
@@ -292,6 +167,7 @@ describe('Blockchains route', async function () {
         .request(app)
         .get('/test/blockchains/active')
         .set('Authorization', `Bearer ${mockedToken}`);
+
       chai.expect(res).to.have.status(200);
       chai.expect(res.body).to.be.an('array');
       const expectedArray = await collectionBlockchains
@@ -306,11 +182,13 @@ describe('Blockchains route', async function () {
       chai.expect(res.body.every((obj) => obj.isActive === true)).to.be.true;
     });
   });
+
   describe('GET blockchain by MongoDBId', async function () {
     it('Should return 403 if no token is provided', async function () {
       const res = await chai.request(app).get('/test/blockchains/1234');
       chai.expect(res).to.have.status(403);
     });
+
     it('Should return blockchain with the proper blockchainId', async function () {
       const createResponse = await createBaseBlockchain(blockchain);
       const res = await chai
@@ -331,169 +209,53 @@ describe('Blockchains route', async function () {
       chai.expect(res).to.have.status(200);
       chai.expect(res.body).to.be.an('object').that.is.empty;
     });
-    testNonMongodbId({
-      method: 'get',
-      path: '/test/blockchains/1111111111111111',
-      body: {},
-      query: {},
-      field: 'blockchainId',
-    });
   });
+
   describe('PUT blockchain', async function () {
-    describe('Core of the route', async function () {
-      it('Should return 403 if no token is provided', async function () {
-        const res = await chai.request(app).put('/test/blockchains/1234');
-        chai.expect(res).to.have.status(403);
-      });
-      modifyBlockchainField({ field: 'chainId', value: 'modifiedChainId' });
-      modifyBlockchainField({ field: 'caipId', value: 'eip155:343' });
-      modifyBlockchainField({ field: 'label', value: 'newLabel' });
-      modifyBlockchainField({ field: 'icon', value: 'newIcon' });
-      modifyBlockchainField({
-        field: 'rpc',
-        value: [
-          'https://new.goerli.blockpi.network/v1/rpc/public',
-          'https://new.rpc.ankr.com/eth_goerli',
-          'https://new.eth-goerli.public.blastapi.io',
-        ],
-      });
-      modifyBlockchainField({
-        field: 'nativeTokenSymbol',
-        value: 'newNativeTokenSymbol',
-      });
-      modifyBlockchainField({ field: 'isEvm', value: true });
-      modifyBlockchainField({
-        field: 'isTestnet',
-        value: true,
-      });
-      modifyBlockchainField({ field: 'isActive', value: true });
-      modifyBlockchainField({
-        field: 'transactionExplorerUrl',
-        value: 'https://new.goerli.etherscan.io/tx/{hash}',
-      });
-      modifyBlockchainField({
-        field: 'addressExplorerUrl',
-        value: 'https://new.goerli.etherscan.io/address/{hash}',
-      });
-      it('Should fail if no blockchain found', async function () {
-        const res = await chai
-          .request(app)
-          .put('/test/blockchains/111111111111111111111111')
-          .set('Authorization', `Bearer ${mockedToken}`);
-        chai.expect(res).to.have.status(404);
-        chai.expect(res.body).to.deep.equal({ msg: 'No blockchain found' });
-      });
+    it('Should return 403 if no token is provided', async function () {
+      const res = await chai.request(app).put('/test/blockchains/1234');
+      chai.expect(res).to.have.status(403);
     });
-    describe('Validators', async function () {
-      it('PUT /blockchains/11111111111111111111 - rpc - Should fail if rpc is not an array', async function () {
-        // Make a request to create the offer with invalid data
-        const res = await chai
-          .request(app)
-          .put('/test/blockchains/11111111111111111111')
-          .set({ Authorization: `Bearer ${mockedToken}` })
-          .send({ rpc: 'notAnArray' });
-        // Assertions
-        chai.expect(res).to.have.status(400);
-        chai.expect(res.body).to.be.an('array');
-        chai.expect(
-          res.body.some(
-            (err) => err.msg === 'must be an array' && err.param === 'rpc'
-          )
-        ).to.be.true;
-      });
-      it('PUT /blockchains/11111111111111111111 - rpc - Should fail if rpc contains non-URL values', async function () {
-        // Make a request to create the offer with invalid data
-        const res = await chai
-          .request(app)
-          .put('/test/blockchains/11111111111111111111')
-          .set({ Authorization: `Bearer ${mockedToken}` })
-          .send({ rpc: ['notAnURL', 123] });
-        // Assertions
-        chai.expect(res).to.have.status(400);
-        chai.expect(res.body).to.be.an('array');
-        chai.expect(
-          res.body.some((err) => err.msg === 'must be an array of URL')
-        ).to.be.true;
-      });
-      const testCases = [
-        'chainId',
-        'caipId',
-        'label',
-        'icon',
-        'rpc',
-        'nativeTokenSymbol',
-        'isEvm',
-        'isTestnet',
-        'isActive',
-        'transactionExplorerUrl',
-        'addressExplorerUrl',
-      ];
-      for (const testCase of testCases) {
-        if (testCase == 'caipId') {
-          testNonCaipId({
-            method: 'put',
-            path: '/test/blockchains/11111111111111111111',
-            body: {
-              ...blockchain,
-              [testCase]: 123,
-            },
-            query: {},
-            field: testCase,
-          });
-        } else if (
-          testCase == 'transactionExplorerUrl' ||
-          testCase == 'addressExplorerUrl'
-        ) {
-          testNonURL({
-            method: 'put',
-            path: '/test/blockchains/11111111111111111111',
-            body: {
-              ...blockchain,
-              [testCase]: 123,
-            },
-            query: {},
-            field: testCase,
-          });
-        } else if (
-          testCase == 'isEvm' ||
-          testCase == 'isTestnet' ||
-          testCase == 'isActive'
-        ) {
-          testNonBoolean({
-            method: 'put',
-            path: '/test/blockchains/11111111111111111111',
-            body: {
-              ...blockchain,
-              [testCase]: 123,
-            },
-            query: {},
-            field: testCase,
-          });
-        } else if (testCase !== 'rpc') {
-          testNonString({
-            method: 'put',
-            path: '/test/blockchains/11111111111111111111',
-            body: {
-              ...blockchain,
-              [testCase]: 123,
-            },
-            query: {},
-            field: testCase,
-          });
-        }
-        testNonEmpty({
-          method: 'put',
-          path: '/test/blockchains/11111111111111111111',
-          body: {
-            ...blockchain,
-            [testCase]: '',
-          },
-          query: {},
-          field: testCase,
-        });
-      }
+    modifyBlockchainField({ field: 'chainId', value: 'modifiedChainId' });
+    modifyBlockchainField({ field: 'caipId', value: 'eip155:343' });
+    modifyBlockchainField({ field: 'label', value: 'newLabel' });
+    modifyBlockchainField({ field: 'icon', value: 'newIcon' });
+    modifyBlockchainField({
+      field: 'rpc',
+      value: [
+        'https://new.goerli.blockpi.network/v1/rpc/public',
+        'https://new.rpc.ankr.com/eth_goerli',
+        'https://new.eth-goerli.public.blastapi.io',
+      ],
+    });
+    modifyBlockchainField({
+      field: 'nativeTokenSymbol',
+      value: 'newNativeTokenSymbol',
+    });
+    modifyBlockchainField({ field: 'isEvm', value: true });
+    modifyBlockchainField({
+      field: 'isTestnet',
+      value: true,
+    });
+    modifyBlockchainField({ field: 'isActive', value: true });
+    modifyBlockchainField({
+      field: 'transactionExplorerUrl',
+      value: 'https://new.goerli.etherscan.io/tx/{hash}',
+    });
+    modifyBlockchainField({
+      field: 'addressExplorerUrl',
+      value: 'https://new.goerli.etherscan.io/address/{hash}',
+    });
+    it('Should fail if no blockchain found', async function () {
+      const res = await chai
+        .request(app)
+        .put('/test/blockchains/111111111111111111111111')
+        .set('Authorization', `Bearer ${mockedToken}`);
+      chai.expect(res).to.have.status(404);
+      chai.expect(res.body).to.deep.equal({ msg: 'No blockchain found' });
     });
   });
+
   describe('DELETE blockchain', async function () {
     it('Should return 403 if no token is provided', async function () {
       const res = await chai
@@ -507,10 +269,7 @@ describe('Blockchains route', async function () {
         .post(pathBlockchains)
         .set('Authorization', `Bearer ${mockedToken}`)
         .send(blockchain);
-      toDeleteDb.push({
-        collection: collectionBlockchains,
-        id: createResponse.body.insertedId,
-      });
+
       chai.expect(createResponse).to.have.status(200);
       chai.expect(createResponse.body).to.have.property('acknowledged', true);
       chai.expect(createResponse.body).to.have.property('insertedId');
@@ -528,186 +287,137 @@ describe('Blockchains route', async function () {
         })
       ).to.be.null;
     });
-    testNonMongodbId({
-      method: 'delete',
-      path: '/test/blockchains/1111111111111111',
-      body: {},
-      query: {},
-      field: 'blockchainId',
-    });
   });
+
   describe('POST useful address', async function () {
-    describe('Core of the route', async function () {
-      it('Should return 403 if no token is not provided', async function () {
-        const res = await chai
-          .request(app)
-          .put('/test/blockchains/useful-address/1234');
-        chai.expect(res).to.have.status(403);
-      });
-      it('Should return 404 if blockchain is not found', async function () {
-        const response = await chai
-          .request(app)
-          .put(`/test/blockchains/useful-address/111111111111111111111111`)
-          .set('Authorization', `Bearer ${mockedToken}`)
-          .send(usefulAddress);
-        chai.expect(response).to.have.status(404);
-        chai.expect(response.body).to.deep.equal({
-          msg: 'No blockchain found',
-        });
-      });
-      it('Should create a new useful address with the proper fields', async function () {
-        const createResponse = await createBaseBlockchain(blockchain);
-        const newUsefulAddress = await chai
-          .request(app)
-          .put(
-            `/test/blockchains/useful-address/${createResponse.body.insertedId}`
-          )
-          .set('Authorization', `Bearer ${mockedToken}`)
-          .send(usefulAddress);
-        chai.expect(newUsefulAddress).to.have.status(200);
-        chai.expect(newUsefulAddress.body).to.deep.equal({
-          acknowledged: true,
-          modifiedCount: 1,
-          upsertedId: null,
-          upsertedCount: 0,
-          matchedCount: 1,
-        });
-        const res = await chai
-          .request(app)
-          .get(`/test/blockchains/${createResponse.body.insertedId}`)
-          .set('Authorization', `Bearer ${mockedToken}`);
-        chai.expect(res).to.have.status(200);
-        delete res.body._id;
-        chai.expect(res.body.usefulAddresses).to.deep.equal({
-          ...blockchain.usefulAddresses,
-          [usefulAddress.contract]: usefulAddress.address,
-        });
-      });
-      it('Should update a useful address with the proper fields', async function () {
-        const createResponse = await createBaseBlockchain(blockchain);
-        const updateUsefulAddress = await chai
-          .request(app)
-          .put(
-            `/test/blockchains/useful-address/${createResponse.body.insertedId}`
-          )
-          .set('Authorization', `Bearer ${mockedToken}`)
-          .send({
-            contract: blockchain.usefulAddresses.myUsefulAddress1,
-            address: 'modifiedAddress',
-          });
-        chai.expect(updateUsefulAddress).to.have.status(200);
-        chai.expect(updateUsefulAddress.body).to.deep.equal({
-          acknowledged: true,
-          modifiedCount: 1,
-          upsertedId: null,
-          upsertedCount: 0,
-          matchedCount: 1,
-        });
-        const res = await chai
-          .request(app)
-          .get(`/test/blockchains/${createResponse.body.insertedId}`)
-          .set('Authorization', `Bearer ${mockedToken}`);
-        chai.expect(res).to.have.status(200);
-        delete res.body._id;
-        chai.expect(res.body.usefulAddresses).to.deep.equal({
-          ...blockchain.usefulAddresses,
-          [blockchain.usefulAddresses.myUsefulAddress1]: 'modifiedAddress',
-        });
+    it('Should return 403 if no token is not provided', async function () {
+      const res = await chai
+        .request(app)
+        .put('/test/blockchains/useful-address/1234');
+      chai.expect(res).to.have.status(403);
+    });
+
+    it('Should return 404 if blockchain is not found', async function () {
+      const response = await chai
+        .request(app)
+        .put(`/test/blockchains/useful-address/111111111111111111111111`)
+        .set('Authorization', `Bearer ${mockedToken}`)
+        .send(usefulAddress);
+      chai.expect(response).to.have.status(404);
+      chai.expect(response.body).to.deep.equal({
+        msg: 'No blockchain found',
       });
     });
-    describe('Validators', async function () {
-      const testCases = ['contract', 'address'];
-      for (const testCase of testCases) {
-        testNonString({
-          method: 'put',
-          path: '/test/blockchains/useful-address/111111111111111111111111',
-          body: {
-            ...usefulAddress,
-            [testCase]: 123,
-          },
-          query: {},
-          field: testCase,
+
+    it('Should create a new useful address with the proper fields', async function () {
+      const createResponse = await createBaseBlockchain(blockchain);
+      const newUsefulAddress = await chai
+        .request(app)
+        .put(
+          `/test/blockchains/useful-address/${createResponse.body.insertedId}`
+        )
+        .set('Authorization', `Bearer ${mockedToken}`)
+        .send(usefulAddress);
+      chai.expect(newUsefulAddress).to.have.status(200);
+      chai.expect(newUsefulAddress.body).to.deep.equal({
+        acknowledged: true,
+        modifiedCount: 1,
+        upsertedId: null,
+        upsertedCount: 0,
+        matchedCount: 1,
+      });
+      const res = await chai
+        .request(app)
+        .get(`/test/blockchains/${createResponse.body.insertedId}`)
+        .set('Authorization', `Bearer ${mockedToken}`);
+      chai.expect(res).to.have.status(200);
+      delete res.body._id;
+      chai.expect(res.body.usefulAddresses).to.deep.equal({
+        ...blockchain.usefulAddresses,
+        [usefulAddress.contract]: usefulAddress.address,
+      });
+    });
+
+    it('Should update a useful address with the proper fields', async function () {
+      const createResponse = await createBaseBlockchain(blockchain);
+      const updateUsefulAddress = await chai
+        .request(app)
+        .put(
+          `/test/blockchains/useful-address/${createResponse.body.insertedId}`
+        )
+        .set('Authorization', `Bearer ${mockedToken}`)
+        .send({
+          contract: blockchain.usefulAddresses.myUsefulAddress1,
+          address: 'modifiedAddress',
         });
-        testNonEmpty({
-          method: 'put',
-          path: '/test/blockchains/useful-address/111111111111111111111111',
-          body: {
-            ...usefulAddress,
-            [testCase]: '',
-          },
-          query: {},
-          field: testCase,
-        });
-      }
+      chai.expect(updateUsefulAddress).to.have.status(200);
+      chai.expect(updateUsefulAddress.body).to.deep.equal({
+        acknowledged: true,
+        modifiedCount: 1,
+        upsertedId: null,
+        upsertedCount: 0,
+        matchedCount: 1,
+      });
+      const res = await chai
+        .request(app)
+        .get(`/test/blockchains/${createResponse.body.insertedId}`)
+        .set('Authorization', `Bearer ${mockedToken}`);
+      chai.expect(res).to.have.status(200);
+      delete res.body._id;
+      chai.expect(res.body.usefulAddresses).to.deep.equal({
+        ...blockchain.usefulAddresses,
+        [blockchain.usefulAddresses.myUsefulAddress1]: 'modifiedAddress',
+      });
     });
   });
+
   describe('DELETE useful address', async function () {
-    describe('Core of the route', async function () {
-      it('Should return 403 if no token is not provided', async function () {
-        const res = await chai
-          .request(app)
-          .delete('/test/blockchains/useful-address/1234');
-        chai.expect(res).to.have.status(403);
-      });
+    it('Should return 403 if no token is not provided', async function () {
+      const res = await chai
+        .request(app)
+        .delete('/test/blockchains/useful-address/1234');
+      chai.expect(res).to.have.status(403);
+    });
 
-      it('Should return 404 if useful address is not found', async function () {
-        const createResponse = await createBaseBlockchain(blockchain);
+    it('Should return 404 if useful address is not found', async function () {
+      const createResponse = await createBaseBlockchain(blockchain);
 
-        const deleteResponse = await chai
-          .request(app)
-          .delete(
-            `/test/blockchains/useful-address/${createResponse.body.insertedId}`
-          )
-          .set('Authorization', `Bearer ${mockedToken}`)
-          .send({ contract: usefulAddress.contract });
-        chai.expect(deleteResponse).to.have.status(404);
-        chai.expect(deleteResponse.body).to.deep.equal({
-          msg: 'No blockchain found or contract doesnt exist',
-        });
-      });
-
-      it('Should return 404 if blockchain is not found', async function () {
-        const deleteResponse = await chai
-          .request(app)
-          .delete('/test/blockchains/useful-address/111111111111111111111111')
-          .set('Authorization', `Bearer ${mockedToken}`)
-          .send({ contract: blockchain.usefulAddresses.myUsefulAddress1 });
-        chai.expect(deleteResponse).to.have.status(404);
-        chai.expect(deleteResponse.body).to.deep.equal({
-          msg: 'No blockchain found or contract doesnt exist',
-        });
-      });
-
-      it('Should delete useful address', async function () {
-        const createResponse = await createBaseBlockchain(blockchain);
-
-        const deleteResponse = await chai
-          .request(app)
-          .delete(
-            `/test/blockchains/useful-address/${createResponse.body.insertedId}`
-          )
-          .set('Authorization', `Bearer ${mockedToken}`)
-          .send({ contract: Object.keys(blockchain.usefulAddresses)[0] });
-        chai.expect(deleteResponse).to.have.status(200);
+      const deleteResponse = await chai
+        .request(app)
+        .delete(
+          `/test/blockchains/useful-address/${createResponse.body.insertedId}`
+        )
+        .set('Authorization', `Bearer ${mockedToken}`)
+        .send({ contract: usefulAddress.contract });
+      chai.expect(deleteResponse).to.have.status(404);
+      chai.expect(deleteResponse.body).to.deep.equal({
+        msg: 'No blockchain found or contract doesnt exist',
       });
     });
 
-    describe('Validators', async function () {
-      testNonString({
-        method: 'delete',
-        path: '/test/blockchains/useful-address/111111111111111111111111',
-        body: { contract: 123 },
-        query: {},
-        field: 'contract',
+    it('Should return 404 if blockchain is not found', async function () {
+      const deleteResponse = await chai
+        .request(app)
+        .delete('/test/blockchains/useful-address/111111111111111111111111')
+        .set('Authorization', `Bearer ${mockedToken}`)
+        .send({ contract: blockchain.usefulAddresses.myUsefulAddress1 });
+      chai.expect(deleteResponse).to.have.status(404);
+      chai.expect(deleteResponse.body).to.deep.equal({
+        msg: 'No blockchain found or contract doesnt exist',
       });
+    });
 
-      testNonEmpty({
-        method: 'delete',
-        path: '/test/blockchains/useful-address/111111111111111111111111',
-        body: { contract: '' },
-        query: {},
-        field: 'contract',
-      });
+    it('Should delete useful address', async function () {
+      const createResponse = await createBaseBlockchain(blockchain);
+
+      const deleteResponse = await chai
+        .request(app)
+        .delete(
+          `/test/blockchains/useful-address/${createResponse.body.insertedId}`
+        )
+        .set('Authorization', `Bearer ${mockedToken}`)
+        .send({ contract: Object.keys(blockchain.usefulAddresses)[0] });
+      chai.expect(deleteResponse).to.have.status(200);
     });
   });
 });
